refactor: client module
- move the client and related connection and net modules into client
