+++ src/main.rs
@@ -41,17 +41,15 @@ extern crate toml;
 mod common;
 mod cfgtypes;
 mod client;
-mod connection;
 mod logger;
-mod net;
 mod options;
 mod stats;
 mod codec;
 mod request;
 
 use client::Client;
+use client::net::InternetProtocol;
 use common::*;
-use net::InternetProtocol;
 use request::{config, workload};
 use std::{env, thread};
 
@@ -112,8 +110,8 @@ pub fn main() {
         }
     };
 
-    let internet_protocol = match net::choose_layer_3(matches.opt_present("ipv4"),
-                                                      matches.opt_present("ipv6")) {
+    let internet_protocol = match client::net::choose_layer_3(matches.opt_present("ipv4"),
+                                                              matches.opt_present("ipv6")) {
         Ok(i) => i,
         Err(e) => {
             halt!("{}", e);
